Put atsign-require in __init__ function, hard-code the PkgId
Precompilation seems to turn the UUID into `toplevel`, so it's not found
as a key in the callback dict.

By requiring that the user put the atsign-require into an __init__
function, this package becomes compatible with packages that have to
define their own operations inside an __init__ (without them having
to call `initm`). Moreover, it makes the package considerably simpler
and less magical.