Issue notifications after loading package
Other packages can set up callbacks to be notified about `@require`
blocks. Previously, the notification was issued before Requires
set the `const` for the module name in the package that set up
the `@require` block. This PR changes it so that those notifications
are sent afterwards, thus ensuring that the `const` will be defined
by the time "consumers" get notified.

This is a potentially-breaking change. Since it would cause huge
`[compat]` churn to bump the major version, given the considerations
below I think we can safely use a minor-version bump:

- Requires' README does not promise any particular ordering.

- This ordering seems to make a bit more sense.

- Revise appears to be the only user of this functionality based on
  the lack of any other entries in `notified_pkgs` and a search of
  JuliaHub

- This change fixes https://github.com/timholy/Revise.jl/issues/658.
  The bug was triggered by Revise changes that processed inputs eagerly,
  in response to new locks to make `Base.require` threadsafe in the
  upcoming Julia 1.8 release.